+++ src/app/components/board/board-service/models.ts
@@ -1,0 +1,14 @@
+type Color = 'white' | 'black' | 'orange';
+
+export interface Square {
+  color: Color;
+  isPlayer: boolean;
+}
+
+export type Row = Square[];
+export type Board = Row[];
+
+export interface Coordinate {
+  x: number;
+  y: number;
+}
